refactor: create IntegrationTestCase.php
