+++ tests/Integration/IntegrationTestCase.php
@@ -1,0 +1,61 @@
+<?php
+
+namespace App\Tests\Integration;
+
+use App\Entity\ManiaplanetUser;
+use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\Tools\SchemaTool;
+use org\bovigo\vfs\vfsStream;
+use Symfony\Bundle\FrameworkBundle\KernelBrowser;
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+
+class IntegrationTestCase extends WebTestCase
+{
+
+    protected EntityManagerInterface $entityManager;
+    protected KernelBrowser $client;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->client = static::createClient();
+        $this->entityManager = self::getContainer()->get(EntityManagerInterface::class);
+
+        $this->initDatabase();
+    }
+
+    protected function givenUploadedFiles(array $uploadedFiles)
+    {
+        vfsStream::setup('root', null, ['uploads' => $uploadedFiles]);
+        $_ENV['UPLOAD_DIR'] = vfsStream::url('root/uploads');
+    }
+
+    protected function givenLoggedInTestuser() {
+        $testUser = $this->createTestUser('testuser', 'test@example.com');
+        $this->client->loginUser($testUser);
+    }
+
+    private function createTestUser(string $login, string $email): ManiaplanetUser
+    {
+        $user = new ManiaplanetUser();
+        $user->setLogin($login);
+        $user->setNickname($login);
+        $user->setEmail($email);
+        $user->setRole('ROLE_USER');
+        $user->setEmailSendApprovalNotification(true);
+
+        $this->entityManager->persist($user);
+        $this->entityManager->flush();
+
+        return $user;
+    }
+
+    private function initDatabase(): void
+    {
+        $metaData = $this->entityManager->getMetadataFactory()->getAllMetadata();
+        $schemaTool = new SchemaTool($this->entityManager);
+        $schemaTool->updateSchema($metaData);
+    }
+
+}
